app/requirements.txt
pyyaml
kubernetes
dacite